refactor: move paged_document_view to core/paging/
Shared paging infrastructure belongs in core/, not features/.
