+++ lib/core/widgets/document_view.dart
