+++ lib/features/document_search/cubit/document_search_cubit.dart
@@ -5,8 +5,8 @@ import 'package:paperless_api/paperless_api.dart';
 import 'package:paperless_mobile/core/database/tables/local_user_app_state.dart';
 import 'package:paperless_mobile/core/notifier/document_changed_notifier.dart';
 import 'package:paperless_mobile/core/service/connectivity_status_service.dart';
-import 'package:paperless_mobile/features/paged_document_view/cubit/document_paging_bloc_mixin.dart';
-import 'package:paperless_mobile/features/paged_document_view/cubit/paged_documents_state.dart';
+import 'package:paperless_mobile/core/paging/cubit/document_paging_bloc_mixin.dart';
+import 'package:paperless_mobile/core/paging/cubit/paged_documents_state.dart';
 import 'package:paperless_mobile/core/model/view_type.dart';
 
 part 'document_search_cubit.g.dart';
