+++ lib/features/documents/cubit/documents_cubit.dart
@@ -6,8 +6,8 @@ import 'package:paperless_mobile/core/database/tables/local_user_app_state.dart'
 import 'package:paperless_mobile/core/extensions/document_extensions.dart';
 import 'package:paperless_mobile/core/notifier/document_changed_notifier.dart';
 import 'package:paperless_mobile/core/service/connectivity_status_service.dart';
-import 'package:paperless_mobile/features/paged_document_view/cubit/document_paging_bloc_mixin.dart';
-import 'package:paperless_mobile/features/paged_document_view/cubit/paged_documents_state.dart';
+import 'package:paperless_mobile/core/paging/cubit/document_paging_bloc_mixin.dart';
+import 'package:paperless_mobile/core/paging/cubit/paged_documents_state.dart';
 import 'package:paperless_mobile/core/model/view_type.dart';
 
 part 'documents_state.dart';
