+++ lib/features/documents/view/widgets/documents_empty_state.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:paperless_api/paperless_api.dart';
 import 'package:paperless_mobile/core/extensions/flutter_extensions.dart';
-import 'package:paperless_mobile/features/paged_document_view/cubit/paged_documents_state.dart';
+import 'package:paperless_mobile/core/paging/cubit/paged_documents_state.dart';
 import 'package:paperless_mobile/generated/l10n/app_localizations.dart';
 
 class DocumentsEmptyState extends StatelessWidget {
