+++ lib/features/saved_view/cubit/saved_view_details_state.dart
