+++ lib/features/similar_documents/cubit/similar_documents_cubit.dart
@@ -3,8 +3,8 @@ import 'package:paperless_api/paperless_api.dart';
 import 'package:paperless_mobile/core/notifier/document_changed_notifier.dart';
 import 'package:paperless_mobile/core/service/connectivity_status_service.dart';
 import 'package:paperless_mobile/core/logging/logger.dart';
-import 'package:paperless_mobile/features/paged_document_view/cubit/document_paging_bloc_mixin.dart';
-import 'package:paperless_mobile/features/paged_document_view/cubit/paged_documents_state.dart';
+import 'package:paperless_mobile/core/paging/cubit/document_paging_bloc_mixin.dart';
+import 'package:paperless_mobile/core/paging/cubit/paged_documents_state.dart';
 
 part 'similar_documents_state.dart';
